refactor(mux): share boundCaptureOutput between backends
The tmux and zellij copies were byte-identical; hoist into multiplexer.ts
alongside the other shared helpers.
