+++ src/multiplexer.ts
@@ -377,3 +377,33 @@ export function execMuxOrThrow(
     });
   }
 }
+
+/** Upper bound for `execFile` maxBuffer on pane-capture subprocesses. */
+export const MAX_CAPTURE_READ_BYTES = 1024 * 1024;
+
+/** Apply line and byte bounds to captured pane output. */
+export function boundCaptureOutput(
+  output: string,
+  opts: CapturePaneOptions,
+): CapturePaneResult {
+  const maxLines = Math.max(0, Math.floor(opts.maxLines));
+  const maxBytes = Math.max(0, Math.floor(opts.maxBytes));
+  let truncated = false;
+  let bounded = output;
+  const lines = bounded.split("\n");
+  if (maxLines > 0 && lines.length > maxLines) {
+    bounded = lines.slice(-maxLines).join("\n");
+    truncated = true;
+  } else if (maxLines === 0 && bounded.length > 0) {
+    bounded = "";
+    truncated = true;
+  }
+  const bytes = Buffer.byteLength(bounded, "utf8");
+  if (bytes > maxBytes) {
+    bounded = Buffer.from(bounded, "utf8")
+      .subarray(bytes - maxBytes)
+      .toString("utf8");
+    truncated = true;
+  }
+  return { output: bounded, truncated };
+}
